Revert "updated home page"
