+++ infected-cell-portal/src/components/infected-cell-pages/CellxgeneInfoBoxView.jsx
@@ -12,22 +12,12 @@ class CellxgeneInfoBoxView extends Component {
         <Banner
           backgroundUrl={"../../images/opencell_logo.png"}
           mainText="Infected Cell"
-          paragraph="Portal for the infected cell project"
+          paragraph="portal for the infected cell project"
         />
         <div className={cs.content}>
-          <Heading title="ABOUT" />
-          <div className={cs.body}>
-            <p>
-            As the COVID-19 pandemic has shown, infectious diseases and emerging new viruses remain a major 
-            threat to human health. Our goals are to understand virus-host interactions as a guide to 
-            developing antiviral therapeutics, to develop vaccines that can prevent infection, to preemptively 
-            identify emerging viruses, and to disseminate the technologies needed to diagnose and discover the 
-            source of microbial infections in the developing world.
-            </p>
-            
-          </div>
+          <Heading title="Mouse lemur" />
           <InfoBox
-            title={"The Infected Cell"}
+            title={"The infected cell"}
             description={"The Infected Cell, an atlas of viral infection"}
             buttonTitle="Read more about it"
             buttonLink="https://opencell.czbiohub.org/"
